+++ composeApp/src/commonMain/kotlin/hu/asztrikx/workout/presentation/logEdit/LogEditEvent.kt
@@ -4,6 +4,6 @@ import kotlinx.datetime.LocalDate
 
 sealed class LogEditEvent {
 	data class Date(val date: LocalDate): LogEditEvent()
-	data class Category(val count: Float?, val id: Int): LogEditEvent()
+	data class Category(val count: Float?, val id: Long): LogEditEvent()
 	data object Save: LogEditEvent()
 }
